btc_hodl_signal/requirements.txt
requests
pandas
selenium
beautifulsoup4
pillow